+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Move the `alchemy-select` pagination checker initializer to a static function
 * Prevent `alchemy-select` from keeping on loading remote data when last few results were empty
 * Fix pagination breaking re-opening of `alchemy-select` dropdown
+* Fix initial value of a `BelongsTo` `alchemy-select` element not loading
 
 ## 0.1.7 (2022-06-12)